refactor: GameCommand 패키지 이동을 통한 도메인의 인프라 의존성 제거
