+++ src/main/java/application/command/GameCommand.java
@@ -1,6 +1,7 @@
-package domain.game;
+package application.command;
 
 import dao.GameDAO;
+import domain.game.Game;
 
 public interface GameCommand {
     void execute(Game game, GameDAO gameDAO, int gameId);
